+++ client/src/actions/tasks.js
@@ -1,13 +1,17 @@
-
-
-import { FETCH_ALL } from '../constants/actionTypes';
+import { FETCH_ALL, CREATE } from '../constants/actionTypes';
 import * as api from '../api';
+import moment from 'moment';
 
 
 export const getTasks = () => async (dispatch) => {
+    const currentDate = moment()
     try{
         const { data } = await api.fetchTasks();
-        // console.log(data);
+        data.map((task) => {
+            //Calculate time difference 
+            const createdDate = moment(task.createdAt);
+            task['timeRemaining'] =  currentDate.diff(createdDate, 'days');
+        })
         dispatch({ type: FETCH_ALL, payload: data });
     } catch (error) {
         console.log(error);
@@ -15,11 +19,31 @@ export const getTasks = () => async (dispatch) => {
 }
 
 export const getClientTask = () => async (dispatch) => {
+    // const currentDate = new Date();
+    const currentDate = moment()
+    
     try{
         const { data } = await api.fetchClientTasks();
-        // console.log(data);
+        data.map((task) => {
+            //Calculate time difference 
+            const createdDate = moment(task.createdAt);
+            task['timeRemaining'] =  task.timeSpan - currentDate.diff(createdDate, 'days');
+        })
         dispatch({ type: FETCH_ALL, payload: data });
     } catch (error) {
         console.log(error);
+    }
+}
+
+export const createTask = (taskData, history) =>  async(dispatch) => {
+    try {
+        const { data } = await api.createTask(taskData);
+
+        dispatch({ type: CREATE, data });
+
+        history.push('/client');
+    } catch (error) {
+        console.log(error);
+        
     }
 }
